refactor: move charm plugin unit tests
